feature(boost bonus): refine bonus
We're adding a slightly more aggressive bonus for racers that are at a
bigger disadvantage. If the difference between the racer's speed is more
than 75% we're giving the disadvantaged player another 50 boost on top
of the percent based boost. This feels more balanced but may require
more work to refine it further.