feature(boost-balance): add boost balance
This commit includes code to give a slight boost strength edge to the
racer that is slower. This should make them more competitive in races,
but not overwhelmingly so.

This may need more tweaking, or it may need a completely different
approach to support making the races slightly more interesting or
balanced. We'll see.