refactor(text ui): add print_centered function
I needed something that I could use to easily print text to the center
of the screen. While it worked well for most of what I wanted, I am
still eyeballing a few other elements. This is a big improvement,
though, over the code littering I had to one off the centered printing.