chore(jekyll): clean up directory structure
The method I was using to hand render the blog posts made things a bit
funky when it came to maintaining where everything lived. So I moved
things around and created a blog-source directory where the work of the
blog rendering with Jekyll would happen. It's not perfect, but this will
be fine given we're using GitHub pages to host this thing.